+++ src/main/java/moe/yushi/authlibinjector/httpd/LegacySkinAPIFilter.java
@@ -3,7 +3,6 @@ package moe.yushi.authlibinjector.httpd;
 import static java.util.Optional.empty;
 import static java.util.Optional.of;
 import static java.util.Optional.ofNullable;
-import static moe.yushi.authlibinjector.internal.fi.iki.elonen.NanoHTTPD.newFixedLengthResponse;
 import static moe.yushi.authlibinjector.util.IOUtils.asString;
 import static moe.yushi.authlibinjector.util.IOUtils.getURL;
 import static moe.yushi.authlibinjector.util.IOUtils.newUncheckedIOException;
@@ -19,9 +18,9 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import moe.yushi.authlibinjector.YggdrasilConfiguration;
-import moe.yushi.authlibinjector.internal.fi.iki.elonen.NanoHTTPD.IHTTPSession;
-import moe.yushi.authlibinjector.internal.fi.iki.elonen.NanoHTTPD.Response;
-import moe.yushi.authlibinjector.internal.fi.iki.elonen.NanoHTTPD.Response.Status;
+import moe.yushi.authlibinjector.internal.fi.iki.elonen.IHTTPSession;
+import moe.yushi.authlibinjector.internal.fi.iki.elonen.Response;
+import moe.yushi.authlibinjector.internal.fi.iki.elonen.Status;
 import moe.yushi.authlibinjector.internal.org.json.simple.JSONObject;
 import moe.yushi.authlibinjector.util.JsonUtils;
 import moe.yushi.authlibinjector.util.Logging;
@@ -73,11 +72,11 @@ public class LegacySkinAPIFilter implements URLFilter {
 				throw newUncheckedIOException("Failed to retrieve skin from " + url, e);
 			}
 			Logging.HTTPD.info("Retrieved skin for " + username + " from " + url + ", " + data.length + " bytes");
-			return of(newFixedLengthResponse(Status.OK, "image/png", new ByteArrayInputStream(data), data.length));
+			return of(Response.newFixedLength(Status.OK, "image/png", new ByteArrayInputStream(data), data.length));
 
 		} else {
 			Logging.HTTPD.info("No skin is found for " + username);
-			return of(newFixedLengthResponse(Status.NOT_FOUND, null, null));
+			return of(Response.newFixedLength(Status.NOT_FOUND, null, null));
 		}
 	}
