+++ src/main/java/moe/yushi/authlibinjector/internal/fi/iki/elonen/IStatus.java
@@ -1,0 +1,8 @@
+package moe.yushi.authlibinjector.internal.fi.iki.elonen;
+
+public interface IStatus {
+
+	String getDescription();
+
+	int getRequestStatus();
+}
